requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
boilerpy3==1.0.7
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docopt==0.6.2
Events==0.5
faiss-cpu==1.7.4
farm-haystack==1.26.4
fastapi==0.115.5
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
inflect==7.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.145
lazy-imports==0.3.1
MarkupSafe==3.0.2
marshmallow==3.23.1
monotonic==1.6
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
num2words==0.5.13
numpy==1.26.4
openai==1.55.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
posthog==3.7.2
prompthub-py==4.0.0
propcache==0.2.0
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-cache==0.9.8
requests-toolbelt==1.0.0
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sseclient-py==1.8.0
starlette==0.41.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.3
typeguard==4.4.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
yarl==1.18.0
